# KuryeNet/docs/openapi.yaml
openapi: 3.0.3
info:
  title: KuryeNet API
  version: 0.0.1
servers:
  - url: http://localhost:8585
paths:
...
  /api/addresses/getByAddressTitle:
    get:
      summary: getByAddressTitle
      parameters:
        - in: query
          name: addressTitle
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/addresses/getByCity:
    get:
      summary: getByCity
      parameters:
        - in: query
          name: city
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/addresses/getById:
    get:
      summary: getByAddressId
      parameters:
        - in: query
          name: id
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/addresses/getByPhoneNumber:
    get:
      summary: getByPhoneNumber
      parameters:
        - in: query
          name: phoneNumber
          schema: { type: string }
      responses:
        '200': { description: OK }
...
  /api/addresses/getallByPage:
    get:
      summary: getAll
      parameters:
        - in: query
          name: pageNo
          schema: { type: string }
        - in: query
          name: pageSize
          schema: { type: string }
      responses:
        '200': { description: OK }
...
  /api/couriers/endOrder:
    post:
      summary: endOrder
      parameters:
        - in: query
          name: orderId
          schema: { type: string }
      responses:
        '200': { description: OK }
...
  /api/couriers/getByCourierCity:
    get:
      summary: getByCourierAddress_City
      parameters:
        - in: query
          name: city
          schema: { type: string }
      responses:
        '200': { description: OK }
...
  /api/couriers/getByCourierIdWithOrderId:
    get:
      summary: getByCourierIdWithOrderId
      parameters:
        - in: query
          name: orderId
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/couriers/getByCourierIdentityNumber:
    get:
      summary: getByCourierIdentityNumber
      parameters:
        - in: query
          name: courierIdentityNumber
          schema: { type: string }
      responses:
        '200': { description: OK }
...
  /api/couriers/getById:
    get:
      summary: getByCourierId
      parameters:
        - in: query
          name: addressId
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/couriers/getCourierWithOrderDetails:
    get:
      summary: getCourierWithOrderDetails
      parameters:
        - in: query
          name: orderNumber
          schema: { type: string }
      responses:
        '200': { description: OK }
...
  /api/couriers/getallByPage:
    get:
      summary: getAll
      parameters:
        - in: query
          name: pageNo
          schema: { type: string }
        - in: query
          name: pageSize
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/couriers/startOrder:
    post:
      summary: startOrder
      parameters:
        - in: query
          name: orderId
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/couriers/updateCourierCoordinates:
    put:
      summary: updateCourierCoordinates
      parameters:
        - in: query
          name: latitude
          schema: { type: string }
        - in: query
          name: longitude
          schema: { type: string }
      responses:
        '200': { description: OK }
...
  /api/customers/getByAddressId:
    get:
      summary: getByCustomerAddress_AddressId
      parameters:
        - in: query
          name: addressId
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/customers/getByCity:
    get:
      summary: getByCustomerAddress_City
      parameters:
        - in: query
          name: city
          schema: { type: string }
      responses:
        '200': { description: OK }
...
  /api/customers/getByCustomerId:
    get:
      summary: getByCustomerId
      parameters:
        - in: query
          name: customerId
          schema: { type: string }
      responses:
        '200': { description: OK }
  /api/customers/getByCustomerNameAndSurname:
    get:
      summary: getByCustomerNameAndCustomerSurname
      parameters:
        - in: query
          name: customerName
          schema: { type: string }
        - in: query
          name: customerSurname
          schema: { type: string }
      responses:
        '200': { description: OK }
...
  /api/customers/getallByPage:
    get:
      summary: getAll
      parameters:
        - in: query
          name: pageNo
          schema: { type: string }
        - in: query
          name: pageSize
          schema: { type: string }
      responses:
        '200': { description: OK }
...